+++ soridam-auth/src/main/java/sorisoop/soridam/auth/common/CustomAccessDeniedHandler.java
@@ -1,7 +1,5 @@
 package sorisoop.soridam.auth.common;
 
-import java.io.IOException;
-
 import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.web.access.AccessDeniedHandler;
 import org.springframework.stereotype.Component;
@@ -13,7 +11,7 @@ import jakarta.servlet.http.HttpServletResponse;
 public class CustomAccessDeniedHandler implements AccessDeniedHandler {
 	@Override
 	public void handle(HttpServletRequest request, HttpServletResponse response,
-		AccessDeniedException accessDeniedException) throws IOException {
+		AccessDeniedException accessDeniedException) {
 		throw new CustomAccessDeniedException();
 	}
 }
