+++ soridam-auth/src/main/java/sorisoop/soridam/auth/jwt/JwtAuthenticationEntryPoint.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Component;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.extern.slf4j.Slf4j;
-import sorisoop.soridam.auth.common.UnauthorizedException;
+import sorisoop.soridam.domain.user.exception.UnauthorizedException;
 
 @Slf4j
 @Component
