+++ soridam-domain/src/main/java/sorisoop/soridam/domain/user/exception/UserDomainExceptionCode.java
@@ -14,7 +14,9 @@ import sorisoop.soridam.common.exception.ExceptionCode;
 public enum UserDomainExceptionCode implements ExceptionCode {
 	USER_NOT_FOUND(NOT_FOUND, "해당 사용자를 찾을 수 없습니다."),
 	INVALID_PASSWORD(BAD_REQUEST, "비밀번호가 일치하지 않습니다."),
-	INVALID_USER(BAD_REQUEST, "사용자가 일치하지 않습니다");
+	INVALID_USER(BAD_REQUEST, "사용자가 일치하지 않습니다"),
+	UNAUTHORIZED(HttpStatus.UNAUTHORIZED, "유효한 인증 정보가 없습니다.")
+	;
 
 	private final HttpStatus status;
 	private final String message;
